#1142 Cleanup some code in the sample
